pyproject.toml
[project]
name = "bedtime-stories"
version = "0.1.0"
description = "Generates unique bedtime story on each invocation."
authors = [
    {name = "n0noob",email = "neemanoop@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.67.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "edge-tts (>=7.0.0,<8.0.0)"
]

[project.urls]
repository = "https://github.com/n0noob/bedtime-stories"

[project.scripts]
bedtime-stories = "bedtime_stories.cli:run_cli"

[tool.poetry]
packages = [{include = "bedtime_stories", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
